requirements.txt
python-fasthtml
monsterui
openai
